Move parsing to its own object.
The abstracted interface is still there. There's no explicit need to
instantiate a member of the parser class since the interface already
does this for you.

This was done mainly because the free functions were essentially a
class anyway that was manually passing around state. The way this is
done is much more idiomatic and preferable.